Updating compilation under Windows Signed-off-by: Rodrigo Javier Tobar Carrizo <rtobar@csrg.inf.utfsm.cl>

+++ include/common.h
@@ -17,7 +17,7 @@
 
 #define NES_PALETTE_COLORS (64)
 
-#define CYCLES_PER_SCANLINE (341/3.)
+#define CYCLES_PER_SCANLINE ((float)(341/3.))
 
 typedef struct {
 	int fd;
